+++ apps/web/src/app/shared/form/form-field/form-field.component.ts
@@ -1,0 +1,28 @@
+import {
+  Component,
+  ContentChild,
+  ElementRef,
+  Input,
+  AfterContentInit,
+  HostBinding,
+} from '@angular/core';
+import { InputDirective } from '../input.directive';
+
+@Component({
+  selector: 'app-form-field',
+  templateUrl: './form-field.component.html',
+})
+export class FormFieldComponent implements AfterContentInit {
+  @Input() label?: string;
+  @ContentChild(InputDirective, { static: true })
+  inputRef?: InputDirective;
+  labelFor?: string;
+
+  ngAfterContentInit(): void {
+    this.labelFor = this.inputRef?.id;
+  }
+
+  get focused(): boolean {
+    return this.inputRef?.focus ?? false;
+  }
+}